Move static data to fixtures
 - also rename fixture file from "abstractUtxo" to "recovery"

Ticket: BG-22317
